poggies

+++ src/str/XtoX/ft_atol.c
@@ -6,7 +6,7 @@
 /*   By: mmoussou <mmoussou@student.42angoulem      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 22:00:18 by mmoussou          #+#    #+#             */
-/*   Updated: 2024/01/11 18:57:52 by mmoussou         ###   ########.fr       */
+/*   Updated: 2024/01/11 23:48:48 by mmoussou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 long	ft_atol(const char *str)
 {
 	long	r;
-	long	i;
+	int		i;
 	char	s;
 
 	r = 0;
